Added Level Selection Text
I updated "Main Menu.py" so that there is now a flag that is activated once the user presses "Enter/Return" on their keyboard. This flag will make it so that the text for the level selection screen will show up. Up next, I will begin work on drawing the shapes that will be used so that they can be clicked by the user.