Worked on Settings for Main Menu.py
I have fixed the glitch that I was facing earlier. Moreover, I have begun to add the necessary text that will be used for the settings page. Later today, I will be adding the necessary buttons for the volume manipulation for the music that is used in the game. I will not be using a slider since we don't have much time to finish this up, and I believe that radio buttons will be easier to use in this scenario.