Worked on Malware Lesson Program
I added the audio files for the first part of the lesson and implemented them so that they could be heard as part of the lesson. Fixed the button counter that wasn't working. I had accidentally placed the rectangle that is used to detect the collision with the mouse in a position that was too high than that of the image.

Moving on, I will begin to work on adding images to the lesson content, finding a way to make sure that the audio is not constantly repeating, finding a way to render the text slowly like Ajad did for his, and adjusting some of the text on some of the sections so that they can be centered.